% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.tex.R
\name{as.tex}
\alias{as.tex}
\title{Coerce to LaTeX Math}
\usage{
as.tex(x)
}
\arguments{
\item{x}{Object.}
}
\description{
The function uses \code{\link[=as.yac_symbol]{as.yac_symbol()}}
to convert \code{x} to an object of class \code{yac_symbol}
and then converts it to LaTeX math
using \code{\link[Ryacas:tex]{Ryacas::tex()}}.
}
\examples{
(x <- matrix(1:10, ncol = 2))
as.tex(x)
}
\seealso{
Other assert functions: 
\code{\link{as.yac_symbol.mat}()},
\code{\link{as.yac_symbol}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{assert functions}
\keyword{assert}
