% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.yac_symbol.mat.R
\name{as.yac_symbol.mat}
\alias{as.yac_symbol.mat}
\alias{as.ysym.mat}
\title{Coerce to \code{yac_symbol} Matrix}
\usage{
as.yac_symbol.mat(x)

as.ysym.mat(x)
}
\arguments{
\item{x}{Object.}
}
\description{
The function uses \code{\link[=as.yac_symbol]{as.yac_symbol()}}
to convert \code{x} to an object of class \code{yac_symbol} \code{y}.
If \code{y$is_mat = TRUE}, the function returns \code{y}.
If \code{y$is_vec = TRUE}, the function converts \code{y}
to a column vector, that is, an \verb{m by 1} matrix.
}
\examples{
as.yac_symbol.mat(1:5)
(x <- matrix(1:10, ncol = 2))
as.yac_symbol.mat(x)
as.ysym.mat(6:10)
(y <- matrix(11:20, ncol = 2))
as.ysym.mat(y)
}
\seealso{
Other assert functions: 
\code{\link{as.tex}()},
\code{\link{as.yac_symbol}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{assert functions}
\keyword{assert}
