% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.yac_symbol.R
\name{as.yac_symbol}
\alias{as.yac_symbol}
\alias{as.ysym}
\title{Coerce to an Object of Class \code{yac_symbol}}
\usage{
as.yac_symbol(x)

as.ysym(x)
}
\arguments{
\item{x}{Object.}
}
\description{
The function checks if \code{x} is of class \code{yac_symbol}.
If the test is \code{TRUE}, it returns \code{x}.
otherwise, it applies \code{\link[Ryacas:ysym]{Ryacas::ysym()}} to \code{x}
and returns the result.
}
\examples{
x <- 1:5
as.yac_symbol(x)
as.yac_symbol(Ryacas::ysym(x))
y <- 6:10
as.ysym(y)
as.ysym(Ryacas::ysym(y))
}
\seealso{
Other assert functions: 
\code{\link{as.tex}()},
\code{\link{as.yac_symbol.mat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{assert functions}
\keyword{assert}
