% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.yac_symbol.R
\name{is.yac_symbol}
\alias{is.yac_symbol}
\alias{is.ysym}
\title{Test for Object of Class \code{yac_symbol}}
\usage{
is.yac_symbol(x)

is.ysym(x)
}
\arguments{
\item{x}{Object.}
}
\description{
Tests if an \code{R} object \code{x} is of class \code{yac_symbol}.
}
\examples{
x <- 1:5
is.yac_symbol(x)
is.yac_symbol(Ryacas::ysym(x))
y <- 6:10
is.ysym(y)
is.ysym(Ryacas::ysym(y))
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{predicate functions}
\keyword{predicate}
