% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yacR.R
\docType{package}
\name{yacR}
\alias{yacR}
\title{yacR: Utilities for Using Ryacas}
\description{
A collection of utility functions
for using Ryacas.
}
\references{
\href{http://www.yacas.org/}{YACAS (Yet Another Computer Algebra System) - http://www.yacas.org/}

\href{http://r-cas.github.io/ryacas/}{Ryacas - http://r-cas.github.io/ryacas/}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{yacR}
