% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.yac_symbol.R
\name{as.yac_symbol}
\alias{as.yac_symbol}
\alias{as.ysym}
\title{Coerce to an Object of Class \code{yac_symbol}}
\usage{
as.yac_symbol(x)

as.ysym(x)
}
\arguments{
\item{x}{Object.}
}
\description{
The function checks if \code{x} is of class \code{yac_symbol}.
If the test is \code{TRUE}, it returns \code{x}.
otherwise, it applies \code{Ryacas::ysym} to \code{x}
and returns the result.
}
\examples{
x <- 1:5
as.yac_symbol(x)
as.yac_symbol(Ryacas::ysym(x))
}
\author{
Ivan Jacob Agaloos Pesigan
}
