% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exe.R
\name{Exe}
\alias{Exe}
\title{Execute Yacas Expression using Ryacas}
\usage{
Exe(expr, str = TRUE, ysym = TRUE, tex = FALSE, simplify = FALSE)
}
\arguments{
\item{expr}{Character string. \code{Yacas} expresion.}

\item{str}{Logical.
If \code{str = TRUE}, returns symbolic result as string.
If \code{str = FALSE}, returns symbolic result as an \code{R} expression.}

\item{ysym}{Logical.
If \code{ysym = TRUE}, returns symbolic result as \code{yac_symbol}
if \code{str = TRUE}.
If \code{ysym = FALSE}, returns symbolic result as string
if \code{str = TRUE}.}

\item{tex}{Logical.
If \code{tex = TRUE} returns results as latex math
if \code{str = TRUE}.
If \code{tex = FALSE}, returns symbolic result as string
if \code{str = TRUE}.}

\item{simplify}{Logical. Simplify symbolic results.}
}
\description{
Executes \code{Yacas} expressions using \code{Ryacas}
with different output options.
}
\examples{
A <- Ryacas::ysym(
  matrix(c("a", "c", "b", "d"), ncol = 2)
)
expr <- paste0("Determinant(", A, ")")
Exe(expr)
Exe(expr, str = TRUE, ysym = FALSE)
Exe(expr, str = TRUE, tex = TRUE)
Exe(expr, str = FALSE)

a <- 1
b <- 2
c <- 3
d <- 4
eval(Exe(expr, str = FALSE))

# `det.yac_symbol` from Ryacas
Ryacas::det(A)

# `det` from base R

A <- matrix(c(a, c, b, d), ncol = 2)
det(A)
}
\references{
\href{http://www.yacas.org/}{YACAS (Yet Another Computer Algebra System) - http://www.yacas.org/}

\href{http://r-cas.github.io/ryacas/}{Ryacas - http://r-cas.github.io/ryacas/}
}
\author{
Ivan Jacob Agaloos Pesigan
}
