% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exe.R
\name{Exe}
\alias{Exe}
\title{Execute Yacas Expression using Ryacas}
\usage{
Exe(x, R = FALSE, format = "ysym", simplify = FALSE)
}
\arguments{
\item{x}{Character string. \code{Yacas} expression.}

\item{R}{Logical.
If \code{R = TRUE}, returns symbolic result as an \code{R} expression.
If \code{R = FALSE}, returns symbolic result as \code{"ysym"}, \code{"str"},  or \code{"tex"}
depending of \code{format}.}

\item{format}{Character string.
Only used when \code{R = FALSE}.
If \code{format = "ysym"},
returns symbolic result as \code{yac_symbol}.
If \code{format = "str"},
returns symbolic result as a characetr string.
If \code{format = "tex"},
returns symbolic result as LaTeX math.}

\item{simplify}{Logical. Simplify symbolic results.}
}
\description{
Executes \code{Yacas} expressions using \code{Ryacas}
with different output options.
}
\examples{
A <- Ryacas::ysym(
  matrix(c("a", "c", "b", "d"), ncol = 2)
)
x <- paste0("Determinant(", A, ")")
Exe(x, R = FALSE, format = "ysym")
Exe(x, R = FALSE, format = "str")
Exe(x, R = FALSE, format = "tex")
Exe(x, R = TRUE)

a <- 1
b <- 2
c <- 3
d <- 4
eval(Exe(x, R = TRUE))

# `det.yac_symbol` from `Ryacas`
Ryacas::det(A)

# `det` from `base` R

A <- matrix(c(a, c, b, d), ncol = 2)
det(A)
}
\references{
\href{http://www.yacas.org/}{YACAS (Yet Another Computer Algebra System) - http://www.yacas.org/}

\href{http://r-cas.github.io/ryacas/}{Ryacas - http://r-cas.github.io/ryacas/}
}
\author{
Ivan Jacob Agaloos Pesigan
}
