% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.yac_symbol.R
\name{is.yac_symbol}
\alias{is.yac_symbol}
\title{Test for object of class \code{yac_symbol}}
\usage{
is.yac_symbol(x)
}
\arguments{
\item{x}{Object.}
}
\description{
Tests if an \code{R} object \code{x} is of class \code{yac_symbol}.
}
\examples{
x <- 1:5
is.yac_symbol(x)
is.yac_symbol(Ryacas::ysym(x))
}
\author{
Ivan Jacob Agaloos Pesigan
}
